% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S04_Plotting.R
\name{add_axes}
\alias{add_axes}
\title{Add Axes to a Plot}
\usage{
add_axes(
  at,
  labels = NULL,
  side = 1,
  tick = F,
  line = NULL,
  cex = 1.25,
  degrees = NULL,
  xpd = NA,
  adj = NULL,
  ...
)
}
\arguments{
\item{at}{The positions (x or y-axis) at which to
add axis labels.}

\item{labels}{Optional vector matching in length to
\code{at} with user-defined labels.}

\item{side}{The side of the plot at which to add axis
labels, either...
\itemize{
\item 1 = bottom;
\item 2 = left;
\item 3 = top;
\item 4 = right.
}}

\item{tick}{Logical; if \code{TRUE} draws tick marks
at the specified positions.}

\item{line}{The number of lines into the margin at which
the axis line will be drawn. If \code{degrees} does not
equal \code{NULL}, must be specified relative
current plotting region (use \code{\link[graphics:text]{par()$usr}}
to get x and y-axis coordinates for plot region).}

\item{cex}{Size of the text.}

\item{degrees}{Number of degrees to rotate text
(note results in a call to \code{\link[graphics:text]{text()}}
rather than \code{\link[graphics:axis]{axis()}}).}

\item{xpd}{A logical value or \code{NA}. If \code{FALSE},
all plotting is clipped to the plot region. If \code{TRUE},
all plotting is clipped to the figure region, and if
\code{NA}, all plotting is clipped to the device region.}

\item{adj}{...}

\item{...}{Additional parameters to be passed to
either \code{\link[graphics]{axis}} or
\code{\link[graphics]{text}} (if a value is
provided for \code{degrees}).}
}
\description{
Wrapper for a call to either \code{\link[graphics]{axis}}
or \code{\link[graphics]{text}}, used to add axis labels
to an existing plot.
}
\examples{
# Create blank plot
blank_plot()
# Draw boundaries
hv_line(x = 0)
hv_line(y = 0)

# Add axes
add_axes(c(0, .5, 1), c("Start", "Middle", "End"))
add_axes(c(0, .5, 1), c("Bottom", "Middle", "Top"), side = 2)
add_axes(.5, "Title", side = 3, cex = 2)

# Create blank plot with custom margins
par(mar = rep(4, 4))
blank_plot()
# Draw boundaries
hv_line(x = 0:1)
hv_line(y = 0:1)

# Add rotated axes
add_axes(c(0, .5, 1), c("Start", "Middle", "End"),
  degrees = 45
)
add_axes(c(0, .5, 1), c("Start", "Middle", "End"),
  degrees = 45, side = 3
)
add_axes(c(0, .5, 1), c("Bottom", "Middle", "Top"),
  degrees = 45, side = 2
)
add_axes(c(0, .5, 1), c("Bottom", "Middle", "Top"),
  degrees = 45, side = 4
)
}
