% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{add_to_list}
\alias{add_to_list}
\title{Add Element to Existing List}
\usage{
add_to_list(l, name, e)
}
\arguments{
\item{l}{A list.}

\item{name}{A character string, the name of the new
element being added.}

\item{e}{An R object, the new element to add to
the list.}
}
\value{
A list with an additional element \code{e}
with the specified name.
}
\description{
A function that adds a new element to a pre-existing
list. Allows greater compatibility with the pipe operator.
}
\examples{
l <- list() |>
  add_to_list( 'A', 1:3 ) |>
  add_to_list( 'B', list( C = 1, D = 2 ) )
print(l)
}
