% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{align_strings}
\alias{align_strings}
\title{Pad Strings to be the Same Length}
\usage{
align_strings(strings, left = TRUE)
}
\arguments{
\item{strings}{A character vector.}

\item{left}{Logical; if \code{TRUE}
left-aligns strings, otherwise
right-aligns strings.}
}
\value{
A character vector.
}
\description{
Function that pads strings to be the same
length, with padding applied either to the
left or right-hand side.
}
\examples{
# Strings of unequal length
x <- c( "A", "BB", "CCC" )

# Left-aligned
s <- align_strings( x )
message( paste0( s, '\n' ) )

# Right-aligned
s <- align_strings( x, FALSE )
message( paste0( s, '\n' ) )

}
