% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S04_Plotting.R
\name{apply_f_to_plot}
\alias{apply_f_to_plot}
\title{...}
\usage{
apply_f_to_plot(dtf, entries = NULL, f = NULL, ...)
}
\arguments{
\item{dtf}{...}

\item{entries}{...}

\item{f}{...}

\item{...}{...}
}
\description{
...
}
\examples{
# Forthcoming
}
