% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arfpam-package.R
\docType{package}
\name{arfpam-package}
\alias{arfpam}
\alias{arfpam-package}
\title{Assorted R Functions for Processing, Analysis, and Modeling}
\description{
The \pkg{arfpam} package provides assorted
functions to help with a variety of common
tasks when processing, plotting, analyzing,
and modeling data. See
\url{https://rettopnivek.github.io/arfpam/}
for more details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rettopnivek.github.io/arfpam}
}

}
\author{
\strong{Maintainer}: Kevin Potter \email{kevin.w.potter@gmail.com}

}
