% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{assign_by_match}
\alias{assign_by_match}
\title{Create Vector with Values Assigned Based on a List of Matches}
\usage{
assign_by_match(values, matches, default = NA)
}
\arguments{
\item{values}{A vector of values equal in length to
\code{matches}.}

\item{matches}{A list of logical vectors each of size N
(see \code{\link{list_of_matches}}).}

\item{default}{The default value used to initialize the
output vector.}
}
\value{
A vector of size N.
}
\description{
This function takes a list of logical vectors equal to
\code{TRUE} for different subsets of a vector, and
generates a new vector with user-specified values based
on each subset from the list.
}
\examples{
# Use 'iris' species for example
data( "iris" ); dtf <- iris

# Regroup 3 iris species into 2 categories
categories <- list( 'setosa', c( 'versicolor', 'virginica' ) )
lst <- list_of_matches( dtf$Species, categories )
dtf$Category <- assign_by_match(1:2, lst)
aggregate( dtf$Category, list( dtf$Species ), unique )

}
\seealso{
\code{\link{apply_f_to_plot}}
}
