% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05_Statistics.R
\name{bounds}
\alias{bounds}
\title{Lower and Upper Limits for Confidence or Credible Intervals}
\usage{
bounds(width)
}
\arguments{
\item{width}{The width of the interval, a numeric value
between 0 and 1.}
}
\value{
A vector with the lower and upper limits.
}
\description{
Given a desired width, computes the lower and upper limit
for a confidence or credible interval. Can be combined
with functions like \code{\link[stat]{quantile}} or
the quantile functions for assorted probability
distributions( e.g., \code{\link[stats:Normal]{qnorm}},
\code{\link[stats:Beta]{qbeta}}, etc.).
}
\examples{
# Lower and upper limits of 95\% interval
bounds( .95 )
# Example data
x <- rnorm( 100, mean = 100, sd = 15 )
# 95\% confidence interval around mean
mean(x) + qnorm( bounds( .95 ) ) * sem( x )
# Predicted
qnorm( bounds( .95 ), mean = 100, sd = 15/sqrt(100) )
# The 1st and 3rd quartiles
quantile( x, bounds( .5 ) )

}
