% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05_Statistics.R
\name{boxcox_transform}
\alias{boxcox_transform}
\title{Box-Cox Transformations for Linear Model}
\usage{
boxcox_transform(
  x,
  outcome = NULL,
  parameter_grid = seq(-3, 3, 0.01),
  output = TRUE
)
}
\arguments{
\item{x}{Either
\itemize{
\item A numeric vector of positive values;
\item A data frame with an outcome variable and
set of predictors for a linear regression.
}}

\item{outcome}{An optional character string with
the column name for the outcome variable in \code{x}.
Otherwise, the first column in \code{x} is assumed
to be the outcome variable.}

\item{parameter_grid}{Vector specifying the grid of
parameters to explore for maximum likelihood estimation.}

\item{output}{Logical; if \code{TRUE}, returns the
outcome variable following the power transformation.}
}
\value{
Either the maximum likelihood estimate for
the power transformation parameter or a vector
for the outcome variable following the transformation.
}
\description{
Wrapper function to the \code{\link[MASS]{boxcox}}
function; estimates the optimal parameter
for the Box-Cox power transformation and
then applies the appropriate power transformation
to the outcome variable.
}
\details{
The Box-Cox power transformation for a vector of
positive values \eqn{x} and parameter \eqn{\lambda} is:

\deqn{ f(x) = \frac{x^{\lambda} - 1}{\lambda}. }

For \eqn{\lambda = 0}, one simply uses the log transform.
}
\examples{
# Simulate 100 values from the normal distribution
set.seed( 3 ); z <- rnorm( 100 )
# Transform the simulated values
x <- (z*.5 + 1)^(1/.5)
# Maximum likelihood estimate for
# transformation parameter
boxcox_transform( x, output = FALSE )

# Histogram of x, transform of x, and
# original simulated values
layout( rbind( 1, 2, 3) )
hist( x )
hist( boxcox_transform( x ) )
hist( z )

}
