% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S04_Plotting.R
\name{col_to_hex}
\alias{col_to_hex}
\title{Convert Colors to Hex Codes}
\usage{
col_to_hex(col, alpha = 1)
}
\arguments{
\item{col}{A character string corresponding to
a supported color name (e.g., 'blue', 'darkred', etc.).
See \code{\link[graphics]{colors}}.}

\item{alpha}{Degree of transparency from
0 (transparent) to 1 (opaque).}
}
\description{
Convert a color name to a hex color code.
}
\examples{
# Create scatter plot for bivariate normal
blank_plot( c(-4,4), c(-4,4) )
# Draw semi-opaque blue points
points( rnorm(1000), rnorm(1000), pch = 19,
        col = col_to_hex( 'blue', alpha = .3 ) )

}
