% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{column_by_other}
\alias{column_by_other}
\title{Unique Values of one Column by Another}
\usage{
column_by_other(dtf, col1, col2)
}
\arguments{
\item{dtf}{A data frame.}

\item{col1}{The first column (non-standard evaluation
possible).}

\item{col2}{The second column (non-standard evaluation
possible).}
}
\value{
A data frame with the unique values
of the first column associated with the
unique values of the second.
}
\description{
A function that takes two columns in a data frame
and reports the unique values of one column
associated with the unique values of the other
column.
}
\examples{
# Define a data frame
dtf <- data.frame(
  A = c( 1, 1, 2, 2, 3, 3 ),
  B = c( 'A', 'A', 'B', 'B', 'C', 'D' )
)

# Values of column 'B' by values of column 'A'
column_by_other( dtf, A, B )

}
\author{
Kevin Potter
}
