% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{copy_from_source}
\alias{copy_from_source}
\title{Copy Files From Source Folder}
\usage{
copy_from_source(
  source_path = "",
  destination_path = "",
  source_subfolder = "",
  environment_var = "FOLDER_SOURCE"
)
}
\arguments{
\item{source_path}{A character string, the
absolute path to the source folder.}

\item{destination_path}{A character string,
the path to the folder to which files should
be copied - if blank, uses the current
working directory.}

\item{source_subfolder}{An optional character
string, the full or partial name of a
subfolder in source location with the files
to copy.}

\item{environment_var}{An optional character
string, the environmental variable with the
path to the source folder.}
}
\value{
As a side effect copies files to a to
the specified destination folder.
}
\description{
Function to copy files from a subfolder in a
source folder to a new subfolder in a user-defined
source folder in the current directory.
}
