% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S07_Projects_and_Reports.R
\name{create_analysis_project}
\alias{create_analysis_project}
\title{Create Files and Folders for Analysis Project}
\usage{
create_analysis_project(initial = TRUE, work = FALSE, proj = "XXX_AYY")
}
\arguments{
\item{initial}{Logical; if \code{TRUE} creates
folders for source data and R scripts, creates
the \code{'_targets.R'} file, and creates
an example R script for cleaning source data.
Otherwise generates a single template R script.}

\item{work}{Logical; if \code{TRUE} using current
work email rather than personal email.}

\item{proj}{A character string, the project label
to start each function with.}
}
\value{
As a side effect, creates a set of folders
and files in the current working directory.
}
\description{
Function to initialize an analysis project folder,
creating folders and files typical to most
analysis projects.
}
