% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{create_table_of_contents}
\alias{create_table_of_contents}
\title{Create a Table of Contents}
\usage{
create_table_of_contents(file_path)
}
\arguments{
\item{file_path}{An absolute path to the
R script.}
}
\value{
Outputs a message to the R console
window with the table of contents.
}
\description{
Given a path to an R script, creates a
table of contents by searching for
section headers of the form:
\code{#### X) Label ####} where \code{X} is a
numbering scheme (e.g.,1, 2.1, 3.2.1)
with periods denoting nested sub-sections.
}
