% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{create_vector}
\alias{create_vector}
\title{Create Code for Vector}
\usage{
create_vector(
  x,
  inner = c("  \\"", "\\",\\n"),
  outer = c("c(\\n", ")"),
  na.rm = TRUE
)
}
\arguments{
\item{x}{A vector of elements.}

\item{inner}{A character vector with \link{1} the
string to add before each element and \link{2}
the string to add after each element.}

\item{outer}{A character with \link{1} the
string to add before all elements and
\link{2} the string to add after all elements.}
}
\value{
Example code in the console.
}
\description{
Function to extract unique elements of
a vector and output example code to the
console.
}
\examples{
create_vector( LETTERS[1:3] )

}
