% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05_Statistics.R
\name{density_points}
\alias{density_points}
\title{Estimate Densities for Individual Observations}
\usage{
density_points(x, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{...}{Additional arguments to be passed
to the density function.}
}
\value{
A list with...
\itemize{
\item x = the sorted values for the original input;
\item y = the associated empirical densities.
}
}
\description{
Given a vector of values, computes the empirical
density for each observation.
}
\examples{
plot(c(-4,4),c(0,.5),type='n',ylab='Density',xlab='z-scores')
x = rnorm( 100 )
dp = density_points( x )
points( dp$x, dp$y, pch = 19 )

}
