% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{dnr}
\alias{dnr}
\title{Do Not Run Multi-Line Code Segments}
\usage{
dnr(x)
}
\arguments{
\item{x}{Any syntactically valid \code{R} expression.}
}
\description{
Allows a user to write out multiple
lines of R code that will not be run.
}
\examples{
# Will not run
dnr({
  print('Hello')
  print('world')
})

# Can be used to also create
# multi-line commments
dnr("
Here are multiple
lines of text
")

}
