% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S04_Plotting.R
\name{draw_borders_and_labels}
\alias{draw_borders_and_labels}
\title{Add Borders and Labels to an Existing Plot}
\usage{
draw_borders_and_labels(
  xl,
  yl,
  labels = c("X-axis", "Y-axis", "Title", ""),
  sides = 1:2,
  lwd = 2,
  lines = 1.5,
  cex = 1.15
)
}
\arguments{
\item{xl}{The lower and upper boundaries for the
x-axis.}

\item{yl}{The lower and upper boundaries for the
y-axis.}

\item{labels}{A character vector of up to 4 elements, the
labels for the bottom, left, top, and right sides,
respectively. If fewer than 4 elements are given, the
corresponding labels are set to the empty character string.}

\item{sides}{An integer vector of up to four values ranging
from 1 to 4, specifying the sides at which to draw a border.
The values 1, 2, 3, and 4 indicate the bottom, left, top,
and right sides, respectively.}

\item{lwd}{A numeric vector of up to 4 values, the width of the
lines (see \code{\link[graphics]{par}}). Values are recycled
if the length is less than 4.}

\item{cex}{A numeric vector of up to 4 values, the size of the
text for the labels (see \code{\link[graphics]{par}}). Values
are recycled if the length is less than 4.}
}
\value{
Adds borders and labels to an existing figure.
}
\description{
Function to add borders and labels to an existing figure.
}
\examples{
# Create a blank plot
xl <- 0:1; yl <- 0:1
blank_plot( xl, yl )

# Add borders and labels
draw_borders_and_labels( xl, yl )

}
