% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S04_Plotting.R
\name{draw_error_bars}
\alias{draw_error_bars}
\title{Add Error Bars to a Plot}
\usage{
draw_error_bars(
  x,
  lower = NULL,
  upper = NULL,
  arrow = TRUE,
  flip = FALSE,
  length = 0.05,
  code = 3,
  angle = 90,
  border = NA,
  columns = NULL,
  ...
)
}
\arguments{
\item{x}{Either a numeric vector or a data frame.
If a numeric vector, specifies the positions at
which error bars should be drawn.}

\item{lower}{A numeric vector with the lower limits
for error bars.}

\item{upper}{A numeric vector with the upper limits
for error bars.}

\item{arrow}{Logical; if \code{TRUE},
adds error bars using \link[graphics:arrows]{graphics::arrows},
otherwise adds error bars using
\link[graphics:polygon]{graphics::polygon} instead.}

\item{flip}{Logical; if \code{TRUE}, error bars
are drawn at positions on the y-axis with limits
specified over the x-axis; otherwise, error bars
are drawn at positions on the x-axis with limits
specified over the y-axis.}

\item{length}{The length of the arrowhead for the
call to \link[graphics:arrows]{graphics::arrows}.}

\item{code}{Integer controlling whether to draw an
arrowhead at the start (1), end (2), or at both ends (3)
of the line for the call to \link[graphics:arrows]{graphics::arrows}.}

\item{angle}{The angle of the lines creating the arrowhead
for the call to \link[graphics:arrows]{graphics::arrows}. Using
90 degrees results in a flat bar per typical error bars.}

\item{border}{The color of the border for calls to
\link[graphics:polygon]{graphics::polygon}.}

\item{columns}{A character vector, the columns for the
positions and lower/upper limits, respectively, if
\code{x} is a data frame.}

\item{...}{Additional plotting parameters for
either the \link[graphics:arrows]{graphics::arrows} or \link[graphics:polygon]{graphics::polygon}.}
}
\description{
Adds error bars to an existing plot.
}
