% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S04_Plotting.R
\name{draw_lines}
\alias{draw_lines}
\title{Add Lines and Error Bars to an Existing Plot}
\usage{
draw_lines(
  x,
  y = NULL,
  lb = NULL,
  ub = NULL,
  columns = NULL,
  pch = 19,
  cex = 1.25,
  lwd = 2,
  lty = 1,
  arrow = FALSE,
  length = 0.05,
  col = "black",
  col.p = "black",
  col.eb = col_to_hex("grey", 0.3),
  bg = "white",
  border = NA
)
}
\arguments{
\item{x}{Either a numeric vector or a data frame.}

\item{y}{An optional numeric vector matching in length to \code{x}.}

\item{lb}{An optional numeric vector matching in length to \code{x},
specifying the lower bounds for error bars.}

\item{ub}{An optional numeric vector matching in length to \code{x},
specifying the upper bounds for error bars.}

\item{columns}{A character vector of either 2 or 4 elements, the
column names for the x and y-axis values and (optionally) the
column names for the lower and upper bounds of the error bars.}

\item{pch}{The type of point to draw
(see \code{\link[graphics]{par}}).}

\item{cex}{The size of the points to draw
(see \code{\link[graphics]{par}}).}

\item{lwd}{The width of the lines
(see \code{\link[graphics]{par}}).}

\item{lty}{The type of line to draw
(see \code{\link[graphics]{par}}).}

\item{arrow}{Logical; if \code{TRUE} draws individual error bars
while if \code{FALSE} draws a single filled bar.}

\item{length}{The width of the caps on the error bars.}

\item{col}{The color of the lines
(see \code{\link[graphics]{par}}).}

\item{col.p}{The color of the points
(see \code{\link[graphics]{par}}).}

\item{col.eb}{The color of the error bars.}

\item{bg}{The background color of the points
(see \code{\link[graphics]{par}}).}

\item{border}{The color for the border of a single
filled error bar.}
}
\value{
Adds lines and error bars to an existing plot.
}
\description{
Function to add lines and error bars to an
existing plot.
}
\examples{
# Draw a line
blank_plot()
draw_lines( c( 0, .5, 1 ), c( 0, .5, 1 ) )

# Pass points in via data frame
draw_lines(
  data.frame( X = c( .1, .2, .3 ), Y = c( .8, .8, .8 ) ),
  lty = 2, lwd = 3, col.p = 'blue', col = 'blue'
)

# Compute mean, SE, and 95\% CI limits for data set
dtf <- aggregate(
  mtcars$mpg, list( mtcars$cyl, mtcars$am ),
  function(x) c( mean(x), sem(x) )
)
dtf$X <- rep( 1:3, 2 )
dtf$M <- dtf$x[,1];
dtf$LB <- dtf$x[,1] - dtf$x[,2] * 1.96
dtf$UB <- dtf$x[,1] + dtf$x[,2] * 1.96

# Create blank plot
xl <- c( .5, 3.5 ); yl <- c( 5, 35 )
blank_plot( xl, yl )
hv_line( h = yl, l = xl )
hv_line( v = xl, l = yl )

# Draw lines for automatic transmission
draw_lines(
  dtf[1:3,],
  columns = c( 'X', 'M', 'LB', 'UB' )
)

# Draw lines for manual transmission
draw_lines(
  dtf[1:3 + 3,],
  columns = c( 'X', 'M', 'LB', 'UB' ),
  col = 'blue', col.p = 'blue', col.eb = col_to_hex( 'blue', .3 )
)

# Add axes and labels
add_axes( 1:3, dtf$Group.1[1:3] )
mtext( 'Cylinders', side = 1, line = 2, cex = 1.25 )
add_axes( c( 10, 20, 30 ), side = 2 )
mtext( 'MPG', side = 2, line = 2, cex = 1.25 )

# Add legend
legend(
  2.5, 30,
  c( 'Automatic', 'Manual' ),
  fill = c( 'black', 'blue' ),
  bty = 'n'
)

}
