% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{duplicate_wide_to_long}
\alias{duplicate_wide_to_long}
\title{Duplicate Values from Wide-Form to Long-Form Data}
\usage{
duplicate_wide_to_long(wf, lf, x, y, default = NA)
}
\arguments{
\item{wf}{A wide-form data frame.}

\item{lf}{A long-form data frame.}

\item{x}{The column in \code{wf} with values to
duplicate (non-standard evaluation possible).}

\item{y}{The column in both \code{wf} and \code{lf}
over which to repeat values over (non-standard
evaluation possible).}

\item{default}{The value to substitute if
no cases for \code{x} based on \code{y}
are found to duplicate.}
}
\value{
A vector matching in length to the number
of rows of \code{lf} with the values of \code{x}
repeated for each unique case of \code{y}.
}
\description{
Function to duplicate values in column \code{x}
from a wide-form (one row per case) data set \code{wf}
based on a shared column \code{y} with a long-form
(multiple rows per case) data set \code{lf}.
}
\examples{
# Example wide-form data-frame
wf <- data.frame(
  ID = 1:3,
  Value = 4:6
)

# Example long-form data-frame
lf <- data.frame(
  ID = rep( 1:3, each = 3),
  Value = NA
)

# Duplicate values from 'wf' based
# on shared # column 'ID'
lf$Value <- duplicate_wide_to_long( wf, lf, x = Value, y = ID )
print( lf )

}
