% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{empty_list}
\alias{empty_list}
\title{Create an Empty List}
\usage{
empty_list(size, labels = NULL)
}
\arguments{
\item{size}{The number of slots.}

\item{labels}{An optional character
vector (whose length equals
\code{size}) with labels for the
slots.}
}
\value{
A list.
}
\description{
Creates a list with a specified
number of empty slots.
}
\examples{
# An empty list with 3 slots
empty_list(3)

# An empty list with labels
empty_list(3, c("S01", "S02", "S03"))
}
