% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{emptyList}
\alias{emptyList}
\title{Create an Empty List}
\usage{
emptyList(size, labels = NULL)
}
\arguments{
\item{size}{The number of slots.}

\item{labels}{An optional character
vector (whose length equals
\code{size}) with labels for the
slots.}
}
\value{
A list.
}
\description{
Creates a list with a specified
number of empty slots.
}
\examples{
# An empty list with 3 slots
emptyList( 3 )

# An empty list with labels
emptyList( 3, c( 'S01', 'S02', 'S03' ) )

}
