% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{env_path}
\alias{env_path}
\title{...}
\usage{
env_path(
  filename = NULL,
  label = NULL,
  subfolders = "",
  folder_names = c("Data", "Documents", "Figures", "Output", "Source", "R"),
  labels = list(c("Data", "data"), c("Documents", "doc"), c("Figures", "fig"),
    c("Output", "out"), c("Source", "src"), "R")
)
}
\arguments{
\item{x}{...}
}
\value{
...
}
\description{
...
}
\examples{
# Forthcoming
}
