% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S02_Math.R
\name{erf}
\alias{erf}
\title{Error Function}
\usage{
erf(x)
}
\arguments{
\item{x}{A vector of values on the real number line.}
}
\value{
A vector of transformed values based on the error function.
}
\description{
Calculates the error function.
}
\examples{
plot(c(-2, 2), c(-1, 1), type = "n", xlab = "x", ylab = "erf(x)")
x <- seq(-2, 2, length = 100)
abline(h = c(-1, 1), col = "grey")
lines(x, erf(x))
}
