% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S04_Plotting.R
\name{error_bars}
\alias{error_bars}
\title{...}
\usage{
error_bars(pos, limits = NULL, lb = NULL, ub = NULL, arrow = T, flip = F, ...)
}
\arguments{
\item{x}{...}

\item{categories}{...}

\item{column}{}
}
\value{
...
}
\description{
...
}
\examples{
# Forthcoming
}
