% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{every}
\alias{every}
\alias{every<-}
\title{Sequence of Values in Regular Increments}
\usage{
every(x, step = 2, start = 1)

every(x, step = 2, start = 1) <- value
}
\arguments{
\item{x}{A vector of values.}

\item{step}{The size of the increment between
indices in the sequence.}

\item{start}{The index at which to start the
sequence.}

\item{value}{A vector of new values to assign
to \code{x} at the sequence of indices.}
}
\value{
A vector of values extracted from \code{x}.
}
\description{
Extracts a sequence of values from a
vector in regular increments.
}
\examples{
# Extract every other value
# at odd positions
every(1:10)
# Extract every other value
# at even positions
every(1:10, , 2) # Note double commas

# Extract every 3rd value starting
# from 6th position
every(1:12, 3, 6)

# Replace values at even
# positions with 0
x <- 1:10
every(x, , 2) <- 0
x
}
