% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{find_filename}
\alias{find_filename}
\title{Check if a File Name can be Found}
\usage{
find_filename(string, output = "logical")
}
\arguments{
\item{string}{A character string, used
for partial string matching.}

\item{output}{The type of output to return.
Options include...
\itemize{
\item \code{'logical'};
\item \code{'vector'};
\item \code{'index'};
\item \code{'name'}.
}}
}
\value{
Either...
\itemize{
\item A single logical value, \code{TRUE} if
any matching file names are found
(\code{'logical'}).
\item A logical vector, \code{TRUE} for all
matching files in the vector of file
names (\code{'vector'}).
\item An integer vector giving the position of
any matches in the vector of file names
(\code{'index'}).
\item A character vector with any matching
file names, otherwise \code{NULL}
(\code{'name'}).
}
}
\description{
Checks if a file name can be found
in a folder via partial string matching.
Multiple types of output are supported.
}
\examples{
# Forthcoming
setwd( find.package( 'arfpam' )[1] )
setwd( 'html' )

find_filename( 'every' )
find_filename( 'every', output = 'vector' )
find_filename( 'every', output = 'index' )
find_filename( 'every', output = 'name' )

}
