% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{find_increment}
\alias{find_increment}
\title{Find Increment Over Range of Values}
\usage{
find_increment(x, n = NULL)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{n}{An integer, the divisor. If not specified,
uses the number of standard deviations instead.}
}
\value{
A named numeric value, the rounded increment
to iterate over the specified number of times,
with the estimated place to round to as a name.
}
\description{
Given a range of values and a desired divisor,
determines the rounded increment to use.
Useful, for example, to determine the
equally-spaced intervals to use for
a figure's axes.
}
\examples{
x <- rnorm(100)
find_increment(x)
find_increment(x, 6)

}
