% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{find_pattern}
\alias{find_pattern}
\title{Find Pattern in a String}
\usage{
find_pattern(s, pattern)
}
\arguments{
\item{s}{A character vector.}
}
\value{
A logical vector.
}
\description{
Function to find a pattern in a vector of
strings via a call to \link[base:grep]{base::grepl}.
}
\examples{
find_pattern( 'a', c( 'cat', 'dog', 'bat' ) )

}
