% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{format_numbers}
\alias{format_numbers}
\title{Pad Numeric Values to be the Same Length}
\usage{
format_numbers(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A character vector.
}
\description{
Function that pads numeric values to be
the same length by adding spaces to the
left-hand side and trailing zeros after
the decimal place.
}
\examples{
# Decimal values
message( paste0( format_numbers( round( rnorm( 3 ), 2 ) ), "\n" ) )
# Whole numbers
message( paste0( format_numbers( rbinom( 3, 100, .1 ) ), "\n" ) )

}
