% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S01_Operators.R
\name{\%btw\%}
\alias{\%btw\%}
\title{Operator to Determine Range Between Which a Value Falls}
\usage{
x \%btw\% y
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{Either a numeric vector specifying a lower and upper
limit, or a character string in the form \code{'(x,y)'},
\code{'[x,y]'}, \code{'(x,y]'}, or \code{'[x,y)'}.
Here \code{x} and \code{y} denote the lower and upper
limits, while square brackets indicate the end point
is included, while round parentheses indicate the end
point is excluded.}
}
\value{
A logical vector.
}
\description{
The binary operator \%btw\% returns a logical vector
indicating all values in a vector that fall within
a specified lower and upper limit.
}
\examples{
x <- 1:10

# Limits specified as numeric vector
x[ x \%btw\% c( 3, 6 ) ]

# Limits specified as character string
x[ x \%btw\% '(3,6)' ]
x[ x \%btw\% '[3,6]' ]
x[ x \%btw\% '(3,6]' ]
x[ x \%btw\% '[3,6)' ]

}
