% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S01_Operators.R
\name{\%w\%}
\alias{\%w\%}
\title{Operator to Check if String is Part of Another String}
\usage{
x \%w\% y
}
\arguments{
\item{x}{A character string.}

\item{y}{A character vector.}
}
\value{
A logical vector, \code{TRUE} if \code{x} is
contained in a given element of \code{y}.
}
\description{
The operator \code{\%w\%} checks if the string
on the left-hand side is contained within the
string on the right-hand side.
}
\details{

}
\examples{
# Check if a string is part of
# another string
"A" \%w\% "ABC"
"D" \%w\% "ABC"

# Vectorized for y
"A" \%w\% c( "ABC", "DEF", "GHI" )

}
