% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{group_index}
\alias{group_index}
\title{Create Index Over Groupings}
\usage{
group_index(..., levels = NULL)
}
\arguments{
\item{...}{Vectors of equal length.}

\item{levels}{A list with the order of the unique levels for
each input vector (indices assigned from first to last level).}
}
\value{
An integer vector from 1 to the number of unique levels.
}
\description{
Create a numeric index over the unique levels of a
variable or a set of variables.
}
\examples{
# Convert to numeric index
group_index( rep( LETTERS[3:1], each = 3 ) )

# Can control assignment of indices
group_index(
  rep( LETTERS[3:1], each = 3 ), levels = list( c( 'C', 'B', 'A' ) )
)

# Can create single index over all
# unique combinations of multiple variables
group_index( rep( LETTERS[1:3], each = 3 ), rep( 1:3, 3 ) )

}
