% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{has_NA}
\alias{has_NA}
\title{Identify NA Values by Row}
\usage{
has_NA(x, any = TRUE)
}
\arguments{
\item{x}{A matrix or data frame.}

\item{any}{Logical; if \code{TRUE} check
if any observation in the row is a \code{NA}
value, else checks if all observations are
\code{NA}.}
}
\value{
A logical vector with values of
\code{TRUE} for rows with any \code{NA}
values (or rows where all values are \code{NA}
when \code{any = FALSE}).
}
\description{
Identifies rows in a matrix or data frame
that contain any (or all) NA values.
}
\examples{
x = matrix( rnorm(9), 3, 3 )
x[2,3] = NA
has_NA( x )
x = data.frame( A = c( 1, 2, NA ), B = 0 )
has_NA( x )

#' x = matrix( rnorm(9), 3, 3 )
x[2,] = NA; x[3,1] = NA
has_NA( x, any = FALSE )

}
