% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S08_Legacy.R
\name{hv_line}
\alias{hv_line}
\title{Draw Horizontal/Vertical Lines}
\usage{
hv_line(h = NULL, v = NULL, l = NULL, ...)
}
\arguments{
\item{h}{A vector with the y-axis positions
for horizontal lines.}

\item{v}{A vector with the x-axis positions
for vertical lines.}

\item{l}{The lower and upper coordinates to determine
the length of the line (if \code{y} is not \code{NULL},
\code{l} is taken as the x-axis coordinates; if
\code{x} is not \code{NULL}, \code{l} is taken as
the y-axis coordinates).}

\item{...}{Additional arguments to be
passed to the \code{\link[graphics]{segments}}
function.}
}
\description{
Draws horizontal or vertical lines on an
existing figure.
}
\examples{
# Create a blank plot
blank_plot()

# Draw horizontal line
hv_line(h = .5, lty = 2)
# Draw vertical line
hv_line(v = .5, lwd = 2)

# Control width of horizontal line
hv_line(h = .25, l = c(.25, .75), col = "blue")
# Control height of vertical line
hv_line(v = .25, l = c(.25, .75), col = "orange")
}
