% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{index_from_levels}
\alias{index_from_levels}
\title{Create Numeric Index Over Unique Levels}
\usage{
index_from_levels(v, levels = NULL)
}
\arguments{
\item{levels}{An optional vector specifying the
order for which numbers should be assigned
to levels.}

\item{obj_vector}{A vector that can be converted
into a factor.}
}
\value{
An integer vector from 1 to the number of
unique levels.
}
\description{
Create a numeric index over the unique levels
of a variable.
}
\examples{
# Variable with 3 unique levels
v <- rep( LETTERS[3:1], each = 3 )

# Convert to numeric index
index_from_levels( v )

# Can control assignment of indices
index_from_levels( v, levels = LETTERS[1:3] )

# To combine over multiple variables use 'paste0'
v1 <- rep( LETTERS[1:2], each = 4 )
v2 <- rep( letters[1:2], 4 )
index_from_levels( paste0( v1, v2 ) )

}
