% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{lin}
\alias{lin}
\title{Create Linear Sequence of Evenly Spaced Values}
\usage{
lin(start, end, n_intervals)
}
\arguments{
\item{start}{The starting value.}

\item{end}{The final value.}

\item{n_intervals}{The number of evenly spaced intervals.}
}
\value{
A numeric vector.
}
\description{
Generates a sequence of evenly spaced
intervals between a lower and upper limit.
}
\examples{
# Five evenly spaced intervals from 0 to 1
lin(0, 1, 5)
}
