% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{list_of_matches}
\alias{list_of_matches}
\title{...}
\usage{
list_of_matches(x, categories, column = NULL)
}
\arguments{
\item{x}{...}

\item{categories}{...}

\item{column}{}
}
\value{
...
}
\description{
...
}
\examples{
# Forthcoming
}
