% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{load_package}
\alias{load_package}
\title{Convenience Function to Load in R Packages}
\usage{
load_package(pkg = c("dplyr", "arfpam"), install = FALSE, quietly = FALSE)
}
\arguments{
\item{pkg}{A character vector of package names to
load - by default loads the current package
in addition to the package
\code{\link[dplyr:dplyr-package]{dplyr}}.}

\item{install}{Logical; if \code{TRUE} attempts
to install missing packages.}

\item{quietly}{Logical; if \code{TRUE} loads
in packages with no console messages (including
warnings and/or error messages).}
}
\value{
Loads in a set of R packages (see
\code{\link[base]{library}}).
}
\description{
Convenience function to load in a set of R packages -
can be used to first install missing packages before
loading them.
}
