% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{load_R_object}
\alias{load_R_object}
\title{Load in R Objects From .RData Files}
\usage{
load_R_object(path_to_rdata, which_object)
}
\arguments{
\item{path_to_rdata}{A character string,
a path to a .RData file to be passed
on to the \code{\link[base]{load}}}

\item{which_object}{The variable name for
the R object to return from the set of
all variables loaded in from the .RData
file. Uses non-standard evaluation.}
}
\description{
A convenience function that, given a
path to a .RData file, will load in
a specified R object contained within
the file.
}
\author{
Kevin Potter
}
