% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S02_Math.R
\name{log_of_sum_of_exp_x}
\alias{log_of_sum_of_exp_x}
\title{Compute the Log of the Sum of the Exponent of Log-Values}
\usage{
log_of_sum_of_exp_x(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector of log-values.}

\item{na.rm}{Logical; if \code{TRUE} removes \code{NA} values.}
}
\value{
The log of the sum of the exponentiated values of \code{x}.
}
\description{
Function to compute the log of the sum of exponent of
a set of log-values.
}
\details{
It is often useful to take the log of very small
or large values (e.g., likelihoods). Working with log-values
makes exponentiation, multiplication, and division easier.
However, addition and subtraction become harder.

Assume \code{ a = log(a.) } and \code{ b = log(b.) }. The identity
\code{log(a. + b.) = a + log(1 + exp(b - a))} allows us to compute
the log of the sum of the exponentiated values of a and b. This can
be further extended to a series of log-values
\code{x = {x[1], x[2], ..., x[n]}} as
\code{x[1] + log( 1 + exp(x[2] - x[1]) + ... + exp(x[n] - x[1]) )}.
}
\examples{
exp_x <- 1:4
x <- log( exp_x )
log( sum( exp_x) )
log_of_sum_of_exp_x( x )

}
\references{
http://bayesfactor.blogspot.com/2016/05/numerical-pitfalls-in-computing-variance.html

https://stackoverflow.com/questions/65233445/how-to-calculate-sums-in-log-space-without-underflow
}
