% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I02_S002_logistic.R
\encoding{UTF-8}
\name{logistic}
\alias{logistic}
\title{Logistic Function}
\usage{
logistic(x)
}
\arguments{
\item{x}{A set of values (\code{-Inf} ≤ x ≤ \code{Inf}).}
}
\value{
Returns a set of probabilities.
}
\description{
This function applies the logistic function to a set of values.
}
\examples{
logistic( c(0,-2.197225,2.1972255,-Inf,Inf) )

}
