% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{make_file_name}
\alias{make_file_name}
\title{Create Formatted File Name}
\usage{
make_file_name(
  description,
  extension,
  project = NULL,
  script = NULL,
  date_time = NULL,
  remove = FALSE,
  path = ".",
  env_variables = c("ABBR_PROJECT", "ABBR_SCRIPT")
)
}
\arguments{
\item{description}{A character string, a human-readable
description of the file (with words typically separated
by \code{'_'}).}

\item{extension}{A character string, the file extension
(e.g., \code{'R'}, \code{'RData'}) - the period is
added automatically.}

\item{project}{A character string, the abbreviation for the
project name (can be inferred from environmental variables).}

\item{script}{A character string, typically a one-letter
abbreviation followed by a two-digit number indicating
the script or analysis index.}

\item{date_time}{A character string giving the date and time
(usually in the format \code{'YYYY_MM_DD-HH_MM'}) - if
not provided uses the current date and time instead.}

\item{remove}{Logical; if \code{TRUE} removes any existing
files in the working directory with the same
project, script, description, and extension values.}

\item{path}{A character string specifying the subfolder(s)
for the file.}

\item{env_variables}{A character vector specifying the
name of the environmental variables with the
project and script abbreviations.}
}
\value{
A character string.
}
\description{
Creates a file name using the standardized
format 'PROJECT-SXX-Description-YYYY_MM_DD-HH_MM.ext'
where 'PROJECT' is a abbreviation for a project name,
and 'SXX' is a letter followed by a number indicating
the script or analysis index. Both variables can
be automatically specified from existing
environmental variables.
}
\examples{

make_file_name( 'Example', 'R' )
make_file_name( 'Example_2', 'RData', project = 'EX', script = 'S01' )

}
