% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{make_filename}
\alias{make_filename}
\title{...}
\usage{
make_filename(
  description,
  extension,
  tag = NULL,
  number = NULL,
  file_date = NULL,
  additional = NULL,
  remove = FALSE
)
}
\arguments{
\item{description}{...}

\item{extension}{...}

\item{tag}{...}

\item{number}{...}

\item{file_date}{...}

\item{additional}{...}

\item{remove}{...}
}
\value{
...
}
\description{
...
}
\examples{
# Forthcoming

}
