% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{match_rows}
\alias{match_rows}
\title{Match Rows Across Two Data Frames}
\usage{
match_rows(dtf_to, dtf_from)
}
\arguments{
\item{dtf_to}{A data frame with the rows to match over.}

\item{dtf_from}{A data frame with the rows to compared against.}
}
\value{
A list equal in length to the number of rows in
\code{dtf_to} with the row indices from \code{dtf_from} that
matches a given row in \code{dtf_to}. If no matches are found
returns \code{NA}.
}
\description{
Function that returns an index indicating
the row in one data frame that matches the
row in another data frame.
}
\examples{
dtf_example_1 <- data.frame( c( LETTERS[3:1], 'Z' ), V2 = c( 3:1, 26 ) )
dtf_example_2 <- data.frame( LETTERS[1:6], V2 = 1:6 )
match_rows( dtf_example_1, dtf_example_2 )

}
