% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{new_limits}
\alias{new_limits}
\title{Rescale Values to Have New Limits}
\usage{
new_limits(x, lower, upper, ...)
}
\arguments{
\item{x}{A vector of numeric values.}

\item{lower}{The new lower limit or minimum value.}

\item{upper}{The new upper limit or maximum value.}

\item{...}{Additional parameters for the
\code{\link[base:max]{min}} and
\code{\link[base:max]{max}} functions.}
}
\value{
A vector of rescaled values.
}
\description{
A function that rescales a vector of
values to have a new minimum and maximum.
}
\examples{
x <- 1:3
new_limits( x, 0, 1 )

}
