% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{not_NA}
\alias{not_NA}
\title{Return a Logical Vector With NA Values set to FALSE}
\usage{
not_NA(lgc)
}
\arguments{
\item{x}{A logical vector.}
}
\value{
A logical vector with \code{NA} cases
set to \code{FALSE}.
}
\description{
Function that takes a logical vector and ensures
any \code{NA} values are set to \code{FALSE}.
}
\examples{
not_NA( c(TRUE, FALSE, NA) )

}
