% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{nq}
\alias{nq}
\title{No Quotation Mark Character Vector}
\usage{
nq(...)
}
\arguments{
\item{...}{Inputs to be converted to character
strings (no special characters).}
}
\value{
A character vector.
}
\description{
Function that allows specification of a character
vector without using quotation marks.
}
\examples{
nq( Hello, world )

}
