% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{palettes}
\alias{palettes}
\title{Various Color Palettes}
\usage{
palettes(type, index = NULL)
}
\arguments{
\item{type}{...}

\item{index}{...}
}
\value{
...
}
\description{
...
}
\examples{
# Forthcoming

}
