% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{path_from_env_var}
\alias{path_from_env_var}
\title{File/Folder Path From an Environment Variable}
\usage{
path_from_env_var(env_var, file_name = NULL, latest = TRUE)
}
\arguments{
\item{env_var}{A character string, the name for
the environment variable.}

\item{file_name}{A character string, the file name
to locate in the folder associated with the
environment variable.}

\item{latest}{Logical; if \code{TRUE} returns only
the latest version of a file whose name contains
a date.}
}
\value{
A character string.
}
\description{
Returns the associated absolute folder path
stored in an existing environment variable,
or the path to a file located in the associated
folder.
}
