% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{percent_that_sums_to_100}
\alias{percent_that_sums_to_100}
\title{Compute Percentage That Sums to 100\%}
\usage{
percent_that_sums_to_100(x, digits = 1, freq = FALSE)
}
\arguments{
\item{x}{A vector of proportions or frequencies
that are suppose to sum to 100\%.}

\item{digits}{The number of digits to round to.}

\item{freq}{Logical; if \code{TRUE} assumes \code{x}
is a vector of frequencies, otherwise assumes
\code{x} is a vector of proportions.}
}
\value{
A vector of percentages that sum to 100\%.
}
\description{
Function that uses the largest remainder method to
ensure that a set of percentages sum to 100\%
even in the presence of rounding error.
}
\examples{
x <- c( 9990, 5, 5 )
# Convert to percentage and round
p <- round( 100*x/sum(x), 1 )
# No longer sums to 100\% due to rounding error
print( sum(p) )
# Adjust percentages using the
# largest remainder method so
# they sum to 100\%
print( percent_that_sums_to_100( p/100 ) )
# Works with frequencies as well
print( percent_that_sums_to_100( x, freq = TRUE ) )

}
