% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S04_Plotting.R
\name{plot_blank}
\alias{plot_blank}
\title{Generate a Blank Plot}
\usage{
plot_blank(x = c(0, 1), y = c(0, 1), guidelines = FALSE, cex = 0.8)
}
\arguments{
\item{x}{The lower and upper boundaries for the
x-axis.}

\item{y}{The lower and upper boundaries for the
y-axis.}

\item{cex}{Text size to use when marking axis line
positions.}

\item{margins}{Logical; if \code{TRUE} displays
guidelines and details regarding the figure
margins to aid in axis and legend specifications.}
}
\value{
An empty plot.
}
\description{
This function generates a completely blank plot.
}
\examples{
# With margin info
plot_blank(guidelines = TRUE)

}
