% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S04_Plotting.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Wrapper for Plotting Histograms}
\usage{
plot_histogram(
  x,
  breaks = "FD",
  border = "white",
  col = "grey",
  main = "",
  plot = TRUE,
  output = FALSE,
  new = TRUE,
  w = 5,
  h = 5,
  ...
)
}
\arguments{
\item{x}{A vector of numeric values.}

\item{breaks}{The argument controlling breakpoints
passed to the \code{\link[graphics]{hist}} function -
by default uses the Freedman-Diaconis algorithm to
find the optimal number of breakpoints.}

\item{border}{The color of the border around the bars.}

\item{col}{The color used to fill the bars.}

\item{main}{The main title for the figure.}

\item{plot}{Logical; if \code{TRUE} generates a figure.}

\item{output}{Logical; if \code{TRUE} returns the
output from the \code{\link[graphics]{hist}} function.
Set to \code{TRUE} if \code{plot} is \code{FALSE}.}

\item{new}{Logical; if \code{TRUE} generates a new
plotting window via a call to \code{\link[grDevices]{x11}}.}

\item{w}{The width (in inches) of the new plotting
window.}

\item{h}{The height (in inches) of the new plotting
window.}

\item{...}{Additional arguments to pass to the
\code{\link[graphics]{hist}} function.}
}
\value{
If \code{output} is \code{TRUE}, returns a
list with the information used to create the
histogram - see the help page for the
\code{\link[graphics]{hist}} function for more
details.
}
\description{
A convenience function making a call to
\code{\link[graphics]{hist}} with changes to
the default options for parameters for pretty
plotting.
}
\examples{
x <- rnorm( 100 )
plot_histogram( x, new = FALSE )

}
