% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S04_Plotting.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Create a Scatter Plot}
\usage{
plot_scatter(
  x,
  y = NULL,
  columns = NULL,
  labels = c("X", "Y"),
  xlim = NULL,
  ylim = NULL,
  new = TRUE,
  w = 5,
  h = 5,
  cex = c(1, 1, 1),
  line = c(1, -1, -1),
  digits = c(2, 2),
  guidelines = TRUE
)
}
\arguments{
\item{x}{Either a data frame or a numeric vector.}

\item{y}{A numeric vector.}

\item{columns}{A character string of two elements,
the columns to use if \code{x} is a data frame.}

\item{labels}{A character string of two elements,
the x and y-axis labels, respectively.}

\item{xlim}{A numeric vector of two values, the
lower and upper limit for the x-axis (in standard
deviation units).}

\item{ylim}{A numeric vector of two values, the
lower and upper limit for the y-axis (in standard
deviation units).}

\item{new}{A logical value; if \code{TRUE} a new
plotting window is generated.}

\item{w}{A numeric value, the width of the new
plotting window.}

\item{h}{A numeric value, the height of the new
plotting window.}

\item{cex}{A numeric vector of three values, the
size of the text for the labels, results, and
axis values, respectively.}

\item{line}{A numeric vector of three values, the
line position for the labels, results, and
axis values, respectively.}

\item{digits}{A integer vector of two values, the
number of digits to round to for the x and
y-axis values, respectively.}
}
\value{
A scatter plot.
}
\description{
Function to create a scatter plot of two variables.
}
\examples{
# Example data
data("iris")
plot_scatter(
  iris, columns = c( 'Petal.Length', 'Sepal.Length' ),
  new = F, labels = c( 'Petal length', 'Sepal length' )
)

}
