% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S02_Math.R
\name{pow}
\alias{pow}
\title{Raise a Value to a Power}
\usage{
pow(x, a)
}
\arguments{
\item{x}{A numeric vector, the base values.}

\item{a}{A numeric vector, the exponents.}
}
\value{
Returns the result of raising \code{x} to
the power \code{a}.
}
\description{
This function raises a value \code{x} to a power
\code{a}. It is useful for improving the readability
of code, clearly separating the base and exponent
from other aspects of an equation.
}
\examples{
2^4
pow(2, 4)

# Sometimes it can be hard to determine
# what is being raised to a power
x <- 2 * 13^4 / 8
# This function makes it easier to tell
x <- 2 * pow(13, 4) / 8
}
