% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{pull_id}
\alias{pull_id}
\title{Extract Subject IDs From a Data-Frame}
\usage{
pull_id(dtf, subject = TRUE, id = NULL)
}
\arguments{
\item{dtf}{A data frame.}

\item{subject}{Logical; if \code{TRUE}
looks for common variable names for
subject identifiers (IDS.CHR.Subject,
ID, etc.). If \code{FALSE} looks for
common variable names for screening
identifiers (IDS.CHR.Screen, etc.).}

\item{id}{A user-defined variable name
if the column with subject identifiers
is not part of the commonly-used labels.}
}
\value{
A vector of values.
}
\description{
Function to extract the unique values for
subject identifiers in a data frame,
assuming a common variable name
(e.g., \code{IDS.CHR.Subject} or \code{ID}).
}
\examples{
dtf <- data.frame( ID = rep( 1:3, each = 3 ), X = rnorm(9) )
pull_id( dtf )

}
