% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{recode}
\alias{recode}
\title{Recode a Variable}
\usage{
recode(x, values, codes = NULL, default = 0, type = NULL)
}
\arguments{
\item{x}{A vector.}

\item{values}{Either a vector of values to
replace, or a list of vectors for the sets of values
to replace. Values of the form \code{'(x,y)'} or
\code{'[x,y]'} result in assignment based on an interval
from x to y, while values of the form \code{'x,y'}
result in a linear trend ranging from 0 to 1 over the
range x to y.}

\item{codes}{A vector of values, either a single
value or a vector matching in length with \code{to_replace}.}

\item{default}{The default value to return as output
for all cases that are not recoded.}

\item{type}{A character string, either \code{'replace'},
\code{'linear'}, or \code{'interval'}. If not provided
function attempts to determine from input to \code{values}.}
}
\value{
A vector.
}
\description{
Function to flexibly recode values of a variable. Can
match values and replace with new codes, assign codes
based on intervals for a continuous variable, or code
a linear trend given a start and end point.
}
\examples{
# Assigning by matching categories
x <- rep( LETTERS[1:3], each = 3 )
recode( x, list( 'A', c('B', 'C') ) )
recode( x, c( 'A', 'B', 'C' ), c( -1, 0, 1 ) )

# Assigning by intervals
x <- seq( -1.5, 1.5, .5 )
cbind(x, recode( x, c( '(-Inf,0]', '(0,Inf]' ) ) )

# Linear trend
cbind(x, recode( x, '-1,1' ) )

}
