% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{rep_y_by_x}
\alias{rep_y_by_x}
\title{Repeat Column Values by Unique Cases in Another Column}
\usage{
rep_y_by_x(lfd, x, y, extra = NULL, default = NA)
}
\arguments{
\item{lfd}{A long-form data frame.}

\item{x}{The column to repeat values over
(non-standard evaluation possible).}

\item{y}{The column with values to duplicate
(non-standard evaluation possible).}

\item{extra}{A logical vector matching in
length to the number of rows of \code{lfd}
specifying additional cases to match by
when isolating the values of \code{y} to
repeat.}

\item{default}{The value to substitute if
no cases for \code{y} based on \code{extra}
are found to duplicate.}
}
\value{
A vector matching in length to the number
of rows of \code{lfd} with the values of \code{y}
repeated for each unique case of \code{x}.
}
\description{
Function to duplicate values in column \code{y} by
the unique cases in column \code{x}. Useful, for
example, and extracting and duplicating a subject's
baseline values across all time points in a long-form
data frame for a longitudinal study.
}
\examples{
# Example long-form data frame
lfd <- data.frame(
  X = rep( 1:3, each = 3 ),
  Y = 1:9,
  E = rep( 1:3, 3 )
)

# Repeat first value of Y for each value of X
i <- lfd$E == 1
rep_y_by_x( lfd, X, Y, extra = i )
# Repeat last value of Y for each value of X
i <- lfd$E == 3
rep_y_by_x( lfd, X, Y, extra = i )

}
