% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{replace_string}
\alias{replace_string}
\title{Replace String Contents}
\usage{
replace_string(s, to_replace, replace_with = "")
}
\arguments{
\item{s}{A character vector.}

\item{to_replace}{A character vector, the patterns to
match and replace within each string.}

\item{replace_with}{An optional character vector,
either of length one or of matching length to
\code{to_replace}, the patterns to substitute.}
}
\value{
A character string.
}
\description{
Function that replaces a specified pattern found
within a string (or vector of strings) with
a user-specified pattern.
}
\examples{
# Example string
x <- c( 'AA', 'AB', 'AC', 'DD' )
# Remove the letter 'A'
replace_string( x, 'A' )
# Replace the letter 'A' with '1'
replace_string( x, 'A', '1' )
# Replace multiple letters
replace_string( x, c( 'B', 'C' ), c( '1', '2' ) )

}
