% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{runs_in_sequence}
\alias{runs_in_sequence}
\title{Determine Runs in a Sequence}
\usage{
runs_in_sequence(x, codes_for_hit = 1)
}
\arguments{
\item{x}{A vector of values.}

\item{codes_for_run}{A vector of the values
in \code{x} indicating a hit.}
}
\value{
A list with a) the total number of runs, and
b) a matrix with a column for the start position of
each run and a column for the end position of each
run.
}
\description{
Given a sequence of values of which
a subset are dubbed 'hits', determine
the number of runs of hits and the
start and end of each run of hits.
}
\examples{
# Generate a sequence of zeros and ones
x <- rbinom( 10, 1, .5 )
print(x)
# Determine runs of ones
runs_in_sequence( x )

}
