% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S04_Plotting.R
\name{save_pdf}
\alias{save_pdf}
\title{Save PDF file}
\usage{
save_pdf(file_name, fun_plot, dmn = c(5, 5), return_file_name = FALSE, ...)
}
\arguments{
\item{file_name}{A character string, the
desired path to the PNG file.}

\item{fun_plot}{A function that generates
a figure as output.}

\item{dmn}{A numeric vector of two values,
the width and height of the figure
to save (default is in inches).}

\item{return_file_name}{A logical value,
if \code{TRUE} returns the file name
as a character string.}

\item{...}{Additional arguments for the
\code{fun_plot} function.}
}
\value{
A PDF file, and optionally the
file name.
}
\description{
Function that generates a figure and saves it
as a PDF file using common settings.
}
