% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S08_Legacy.R
\name{save_png_figure}
\alias{save_png_figure}
\title{Create and save PNG file}
\usage{
save_png_figure(plotting_fun, file_name, figure_dim = c(5, 5), res = 300, ...)
}
\arguments{
\item{plotting_fun}{A function, with the
creation of a figure as a side effect.}

\item{file_name}{A character string, the
file name for the PNG file - must end
in '.png'.}

\item{figure_dim}{A numeric vector of two
values, the width and height of the
figure in inches.}

\item{res}{A integer value, the resolution
for the PNG file.}

\item{...}{Additional arguments to pass to
\code{plotting_fun}.}
}
\value{
Output from \code{plotting_fun}, if any.
}
\description{
Function to create and save a PNG file
given a plotting function to run.
}
