% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S04_Plotting.R
\name{save_tiff}
\alias{save_tiff}
\title{Save TIFF file}
\usage{
save_tiff(
  file_name,
  fun_plot,
  dmn = c(5, 5),
  units = "in",
  res = 300,
  compression = "none",
  restoreConsole = TRUE,
  pointsize = 12,
  bg = "white",
  family = "",
  type = c("windows", "cairo"),
  symbolfamily = "default",
  return_file_name = FALSE,
  ...
)
}
\arguments{
\item{file_name}{A character string, the
desired path to the PNG file.}

\item{fun_plot}{A function that generates
a figure as output.}

\item{dmn}{A numeric vector of two values,
the width and height of the figure
to save (default is in inches).}

\item{units}{A character string, the
units for figure dimensions (see
\link[grDevices:png]{grDevices::tiff}).}

\item{res}{An integer value, the
figure resolution (see
\link[grDevices:png]{grDevices::tiff}).}

\item{compression}{A character string, the type
of compression to use; see \link[grDevices:png]{grDevices::tiff}).}

\item{restoreConsole}{A logical value, see
\link[grDevices:png]{grDevices::tiff}).}

\item{pointsize}{An integer value, the point
size of text (see \link[grDevices:png]{grDevices::tiff}).}

\item{bg}{a character string, the background
color (see \link[grDevices:png]{grDevices::tiff}).}

\item{family}{A character string specifying the
font family (see \link[grDevices:png]{grDevices::tiff}).}

\item{type}{A character string governing how
plotting is done (see \link[grDevices:png]{grDevices::tiff}).}

\item{symbolfamily}{A character stringth for
cairographics options (see \link[grDevices:png]{grDevices::tiff}).}

\item{return_file_name}{A logical value,
if \code{TRUE} returns the file name
as a character string.}

\item{...}{Additional arguments for the
\code{fun_plot} function.}
}
\value{
A JEPG file, and optionally the
file name.
}
\description{
Function that generates a figure and saves it
as a TIFF file using common settings.
}
