% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{section}
\alias{section}
\title{Section Numbers for Tracking Progress}
\usage{
section(x, run = TRUE, spacing = ".", end = ")")
}
\arguments{
\item{x}{A character string, such as
'1)' or '2.3)' or of a similar form.}

\item{run}{Logical; if TRUE, a message is
generated.}

\item{spacing}{A character string, the
character to use to determine the
number of indents to use to
identify section hierarchies.}

\item{end}{A character string, the
final symbol to attach to the end of
a section header.}
}
\description{
Output section numbers to the console via
a call to \code{\link[base]{message}}.
Useful for tracking progress when running
a script and for debugging where errors
occur.
}
\examples{

section("1")
# Periods indent output by 2 spaces
section("1.1")
section("1.1.1")
}
