% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05_Statistics.R
\name{sem}
\alias{sem}
\title{Standard Error of the Mean}
\usage{
sem(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Logical; if \code{TRUE}, removes \code{NA}
values first.}
}
\value{
The standard error of the mean.
}
\description{
This function calculates the standard error
of the mean for a set of numeric values.
}
\details{
Given the standard deviation \eqn{s_x} and sample size
\eqn{n} of a sample \eqn{x}, the standard error of the mean is:

\deqn{ \frac{s_x}{\sqrt{n}}. }
}
\examples{
# Simulate 100 values from a normal distribution
set.seed(2)
x <- rnorm(100)
# Standard error of the mean should be ~0.1
sem(x)
}
