% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S04_Plotting.R
\name{specify_positions}
\alias{specify_positions}
\title{Specify Positions over Grouping Factors}
\usage{
specify_positions(dtf, spacing = 1)
}
\arguments{
\item{dtf}{A data frame with separate columns for
each grouping variable.}

\item{spacing}{A numeric value, the space to add
between groups.}
}
\value{
A numeric vector of values.
}
\description{
Function to create positions for a figure based
on the combination of grouping factors in a
data frame.
}
\examples{
# Example data set
data("iris")
dtf <- iris
dtf$Long_petals <- dtf$Petal.Length > median( dtf$Petal.Length )
dtf <- stats_by_group( dtf, 'Sepal.Length', c( 'Species', 'Long_petals' ) )
specify_positions( dtf[, 1:2] )

}
