% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{squish}
\alias{squish}
\title{Collapse a Character Vector}
\usage{
squish(chr_vec, between = "")
}
\arguments{
\item{chr_vec}{A character vector.}

\item{between}{The value to use as spacing between
each element in \code{chr_vec}.}
}
\value{
A character string.
}
\description{
Function that combines elements of a
character vector into a single string.
}
\examples{
# Collapse a character vector
print( squish( c( "A", "B", "C" ) ) )

# Collapse a character vector with custom spacing
print( squish( c( "1", "2", "3" ), " + " ) )

}
