% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05_Statistics.R
\name{standardize}
\alias{standardize}
\title{Standardize Columns in a Data Frame or Matrix}
\usage{
standardize(
  x,
  y = NULL,
  mean_sd = NULL,
  raw = FALSE,
  as_list = FALSE,
  labels = c("X", "Y")
)
}
\arguments{
\item{x}{A data frame or matrix of numeric values.}

\item{y}{A data frame or matrix of numeric values
(must have the same column names in same order
as \code{x}).}

\item{mean_sd}{A list of two numeric vectors equal in
length to the number of columns with the means and
standard deviations, respectively, to use for scaling.}

\item{raw}{Logical; if \code{TRUE}, uses the means
and standard deviations given in \code{mean_sd}
to return the original raw values of \code{x}.}

\item{as_list}{Logical; if \code{TRUE} returns
a named list with the scaled values of \code{x}
(and \code{y} if provided) along with the means and
standard deviations used for scaling. Automatically
set to \code{TRUE} when \code{y} is provided.}

\item{labels}{A character vector with the labels for
the \code{x} and \code{y} data sets if returning a
list.}
}
\value{
Either a scaled data frame or matrix or a list
with the scaled values and the means and standard deviations
used for scaling.
}
\description{
Function to standardize (mean-center and scale
by standard deviation resulting in a mean of 0
and standard deviation of 1) columns in a
matrix or data frame.
}
\examples{
# Create data frame
x_raw <- round( matrix( rnorm( 9, 100, 15 ), 3, 3 ) )
colnames(x_raw) <- paste0( 'X', 1:3 )
print(x_raw)

# Standardize columns
x <- standardize( x_raw )
print(x)

# Create second data frame with same
# variables but new values
y_raw <- round( matrix( rnorm( 9, 50, 15 ), 3, 3 ) )
colnames(y_raw) <- paste0( 'X', 1:3 )
print(y_raw)

# Scale columns of y_raw based on means and
# standard deviations from x_raw
lst <- standardize( x_raw, y_raw, labels = c('x', 'y') )
y <- lst$Data$y
print( y )

# Undo scaling
standardize( y, mean_sd = lst$Scaling, raw = TRUE )

}
