% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{start_of_new}
\alias{start_of_new}
\title{Identify Start of New Repetition}
\usage{
start_of_new(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A logical vector equal to \code{TRUE}
for each instance in which a subsequent value
differs from a prior value.
}
\description{
Function that identifies when a new repetition
of values in a vector starts.
}
\examples{
x <- rep( 1:4, each = 3 )
start_of_new( x )

}
