% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05_Statistics.R
\name{statistic}
\alias{statistic}
\title{Compute a Statistic}
\usage{
statistic(
  x,
  f = length,
  include = NULL,
  exclude = NULL,
  na.rm = T,
  default = NA,
  ...
)
}
\arguments{
\item{x}{A vector.}

\item{f}{A function that takes the vector
\code{x} as its first argument.}

\item{include}{A logical vector matching
length to the vector \code{x}.}

\item{exclude}{A vector of unique cases
to match and exclude.}

\item{na.rm}{Logical; if \code{TRUE}
removes \code{NA} values from \code{x}.}

\item{default}{The default value to
return if \code{x} is empty.}

\item{...}{Additional arguments to
pass to the function \code{f}.}
}
\value{
A computed statistic, output from
the user-supplied function.
}
\description{
This function robustly computes a statistic
over a vector of values. The user can
specify what to return if the vector
is empty, values to include or exclude,
and how to treat missing values. Useful in
combination of functions like
\code{\link[stats]{aggregate}} or
\code{\link[dplyr:summarise]{dplyr's summarise}}.
}
\examples{
# Examples using the 'iris' data set
data( "iris" )

# Default
statistic( iris$Sepal.Length )
# User-specified statistic
statistic( iris$Sepal.Length, f = mean )
# Include subset of cases
statistic( iris$Sepal.Length, f = mean,
           include = iris$Species == 'setosa' )
# Custom function
statistic( iris$Species, f = function(x) mean( x \%in\% 'setosa' ) )
# Exclude unique cases
statistic( iris$Species, f = function(x) mean( x \%in\% 'setosa' ),
           exclude = 'virginica' )
# If empty vector supplied, user-specified default value returned
statistic( iris$wrong_name, default = 0 )

}
