% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05_Statistics.R
\name{stats_by_group}
\alias{stats_by_group}
\title{Compute Statistics by Group}
\usage{
stats_by_group(
  dtf,
  column,
  groupings,
  statistics = c("M", "SD"),
  method = "Student's T",
  categories = 1,
  width = 0.95,
  na.rm = TRUE
)
}
\arguments{
\item{dtf}{A data frame.}

\item{column}{A character string, the column in \code{dtf} to
compute statistics for.}

\item{groupings}{A character vector, the columns in \code{dtf}
to use as grouping factors (it is recommended that they all
be categorical variables).}

\item{statistics}{A character vector, the set of different
statistics to compute over groups.}

\item{method}{A character string, the type of method to use
when computing uncertainty intervals. Options include:
\code{"Student's T"} for means or
\code{"Beta-binomial"} for proportions.}

\item{categories}{An optional vector of elements to match over
when computing frequencies, proportions, or percentages.}

\item{width}{A numeric value between 0 and 1, the width for
uncertainty intervals.}

\item{na.rm}{A logical value; if \code{TRUE} removes
\code{NA} values.}
}
\value{
A data frame with separate rows for each combination
of grouping factors and separate columns for each statistic
to compute.
}
\description{
A function to compute assorted univariate statistics
for a specified variable in a data frame over desired
grouping factors.
}
\details{
Possible univariate statistics that can be computed:
\itemize{
\item \code{'N'} = Sample size;
\item \code{'M'} = Mean;
\item \code{'Md'} = Median;
\item \code{'SD'} = Standard deviation;
\item \code{'SE'} = Standard error of the mean;
\item \code{'C'} = Counts/frequencies;
\item \code{'Pr'} = Proportions;
\item \code{'P'} = Percentages.
}

Additionally, specifying \code{'UI'} in combination with the
argument \code{method} will compute the lower and upper limits
of a desired uncertainty interval. The width of the interval
can be controlled by the argument \code{width}.
}
\examples{
# Example data set
data(iris)
dtf <- iris

# Mean/SD for sepal length by species
dtf |> stats_by_group( 'Sepal.Length', 'Species' )

# Create additional categorical variable
dtf$Long_petal <- c( 'No', 'Yes' )[
  ( dtf$Petal.Length > median( dtf$Petal.Length) ) + 1
]
# Sample size, mean, and confidence intervals using Student's T
# distribution by species and whether petals are long
dtf |> stats_by_group(
  'Sepal.Length', c( 'Species', 'Long_petal' ), c( 'N', 'M', 'UI' )
)

# Create additional categorical variable
dtf$Long_sepal <- c( 'No', 'Yes' )[
  ( dtf$Sepal.Length > median( dtf$Sepal.Length) ) + 1
]
# Proportion and confidence intervals based on beta-binomial
# distribution for long sepals by long petals
dtf |> stats_by_group(
  'Long_sepal', c( 'Long_petal' ), c( 'N', 'Pr', 'UI' ),
  categories = 'Yes', method = 'Beta-binomial'
)

}
