% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{strip_value}
\alias{strip_value}
\title{Strip Vector of a Value}
\usage{
strip_value(x, value = NA)
}
\arguments{
\item{x}{A vector of values.}

\item{value}{The value to remove (defaults
to \code{NA})}
}
\value{
A vector of values sans the one removed.
}
\description{
Removes a specified value from a vector.
}
\examples{
x <- c( 1, 2, NA, 3 )
print( strip_value(x) )

x <- c( 'Hello', '', 'world' )
print( strip_value(x, '') )

}
