% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05_Statistics.R
\name{summa}
\alias{summa}
\title{...}
\usage{
summa(
  x,
  syntax = "[[M]] ([[SD]])",
  categories = NULL,
  digits = NULL,
  na.rm = TRUE,
  pad = FALSE,
  f = NULL,
  ...
)
}
\arguments{
\item{x}{...}

\item{syntax}{...}

\item{categories}{...}

\item{digits}{...}

\item{na.rm}{...}

\item{pad}{...}

\item{f}{...}

\item{...}{...}
}
\value{
...
}
\description{
...
}
\details{
...
}
\examples{
# Forthcoming
# Example using 'iris' data set
data("iris")
# Continuous variable - sepal length
x <- iris$Sepal.Length

# Mean and standard deviation
summa(x)
# Median and IQR
summa(x, "[[M]] ([[IQR]])")
# Pad to 2 decimal places
summa(x, "[[M]] ([[IQR]])", pad = TRUE)
# Mean (SD); N [min and max]
summa(x, "[[N]]; [[M]] ([[SD]]); " \%p\%
  "[[[Mn]], [[Q1]], [[Md]], [[Q3]], [[Mx]]]",
digits = 1
)

# Custom measures via user-defined function
# (e.g., bootstrapped confidence interval)
fnc <- function(x) {
  btstrp <- bootstrap(
    x,
    summary = function(y) quantile(y, c(.025, .975))
  )
  return(btstrp$summary)
}
summa(x, "[[M]] ([[SE]]) [[[1]] to [[2]]]",
  f = fnc
)

# Example using 'mtcars' data set
# Categorical variable - # of forward gears
data("mtcars")
x <- mtcars$gear

# Percent and counts for 3 forward gears
summa(x == 3, "[[P]]\% ([[C]] out of [[N]])")
# Percent and counts for 4 or 5 forward gears
summa(x, "[[P]]\% ([[C]] out of [[N]])",
  categories = c(4, 5)
)
}
