% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{take_x_by_y}
\alias{take_x_by_y}
\title{Repeat Column Values by Unique Cases in Another Column}
\usage{
take_x_by_y(lfd, x, y, extra = NULL, default = NA, per_row = TRUE)
}
\arguments{
\item{lfd}{A long-form data frame.}

\item{x}{The column with values to duplicate
(non-standard evaluation possible).}

\item{y}{The column to repeat values over
(non-standard evaluation possible).}

\item{extra}{A logical vector matching in
length to the number of rows of \code{lfd}
specifying additional cases to match by
when isolating the values of \code{x} to
repeat.}

\item{default}{The value to substitute if
no cases for \code{x} based on \code{extra}
are found to duplicate.}

\item{per_row}{Logical; if \code{FALSE} returns
the associated value of \code{x} for each
unique value of \code{y}, otherwise returns
a value of \code{x} per each frow of \code{lfd}.}
}
\value{
A vector matching in length to the number
of rows of \code{lfd} (for \code{per_row = TRUE}) or
to the number of unique values of \code{y} with the
values of \code{x} repeated for each unique case of
\code{y}.
}
\description{
Function to duplicate values in column \code{x} by
the unique cases in column \code{y}. Useful, for
example, in extracting and duplicating a subject's
baseline values across all time points in a long-form
data frame for a longitudinal study.
}
\examples{
# Example long-form data frame
lfd <- data.frame(
  ID = rep( 1:3, each = 3 ),
  Value = 1:9,
  Time = rep( 0:2, 3 )
)

# Repeat first value of Y for each value of X
i <- lfd$Time == 0
take_x_by_y( lfd, Value, ID, extra = i )
# Repeat last value of Y for each value of X
i <- lfd$Time == 2
take_x_by_y( lfd, Value, ID, extra = i )
# Per unique case of ID
take_x_by_y( lfd, Value, ID, extra = i, per_row = FALSE )

}
