% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{templates}
\alias{templates}
\title{Create Templates for Annotations and Code}
\usage{
templates(type, author = "Kevin Potter", email = "kevin.w.potter@gmail.com")
}
\arguments{
\item{type}{...}

\item{author}{...}

\item{email}{...}
}
\value{
...
}
\description{
...
}
\examples{
# Forthcoming

}
