% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{term_coding}
\alias{term_coding}
\alias{term_coding_dummy}
\alias{term_coding_effect}
\title{Parameters for Recoding Variables}
\usage{
term_coding(values, codes = NULL, default = 0, type = NULL)

term_coding_dummy(categories)

term_coding_effect(categories, referent)
}
\arguments{
\item{values}{A vector or list of values. Must
equal in length to \code{codes}.}

\item{codes}{A vector of values.}

\item{type}{A character string, either \code{'replace'},
\code{'linear'}, or \code{'interval'}. If not provided
function attempts to determine from input to \code{values}.}

\item{categories}{A vector of values, all cases
to be coded as 1 (dummy coding and effect coding).}

\item{referent}{A vector of values, all cases
to be coded as -1 (effect coding).}

\item{default.}{A value.}
}
\value{
A list.
}
\description{
Function to help format input to the
\link{term_prep} function. Initializes
a list with the specific parameters to be
passed to the \link{recode} function
when creating a new term.
}
