% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{term_combo}
\alias{term_combo}
\title{Template for Combining Terms}
\usage{
term_combo(combine, label = "", transformation = NULL, scale = NULL)
}
\arguments{
\item{combine}{A character vector, the names
of the terms/variables to combine. Providing a
named vector}

\item{label}{An optional label for the new term.}

\item{transformation}{An optional character string
specifying an R expression indicating how terms
should be combined. Variables are matched to the
named elements of \code{combine}. If not provided
defaults to multiplying all terms together (i.e.,
an interaction).}

\item{scale}{Either a logical value, or a named
vector of the form \code{c(m = ..., sd = ...)}
for standardizing the term.}
}
\value{
A list.
}
\description{
Function to help format input to the
\link{term_prep} function. Initializes
the sublist with the parameters for creating
a new term by combining other existing
terms (e.g., an interaction).
}
