% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{term_prep}
\alias{term_prep}
\title{Create New Terms for Data}
\usage{
term_prep(x, settings, output = "x")
}
\arguments{
\item{x}{A data frame.}

\item{settings}{A named list, in the form:
\code{list(column = list( new1 = list(...), new2 = list(...)))},
where \code{column} is a pre-existing variable in \code{x},
and \code{new1}, \code{new2}, etc. are new terms to be created
using variable \code{column}.}

\item{output}{A character string, the type of output to return,
where \code{'x'} (the default) returns an updated data frame,
\code{'settings'} returns an updated list, and \code{'both'}
returns a list with both the data frame and list of settings.}
}
\value{
Either a data frame, a list of settings, or a list with
both the data frame and list of settings.
}
\description{
Function that takes a list specifying base variables and
new terms to create and updates a data frame appropriately.
Allows for easy pre-processing of multiple data frames in
the same manner for analyses (e.g., when prepping training
and validation sets for data).
}
\examples{
data("mtcars")
# Split into two sets of data
x1 <- mtcars[ seq( 1, 32, 2), ] # Odd
x2 <- mtcars[ seq( 1, 32, 2), ] # Even

lst <- list(
   mpg = list(
     outcome = term_new(
      label = 'Miles per gallon'
    )
  ),
  hp = list(
    log_hp = term_new(
      label = 'Log(Horsepower)',
      transformation = 'log(x)',
      scale = TRUE,
      order = c( 't', 's' )
    )
  ),
  vs = list(
    vs_0v1 = term_new(
      label = 'Engine: V-shaped vs. straight',
      coding = term_coding_effect( 1, 0 ),
      scale = TRUE,
      order = c( 'c', 's' )
    )
  )
)
# Add info on mean/SD from 'x1' data
lst <- x1 |> term_prep( lst, output = 'settings' )
# Update 'x1' and 'x2'
x1 <- x1 |> term_prep( lst )
x2 <- x2 |> term_prep( lst )

# Fit 'x1' data
fit <- lm( outcome ~ log_hp + vs_0v1, data = x1)
# Predict 'x2' data
predict( fit, newdata = x2 )
}
