% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{update_package_version}
\alias{update_package_version}
\title{Update Package Versions in Script Documentation}
\usage{
update_package_version(
  file_name,
  prior_text_package = "#   * The package ",
  prior_text_version = " (version ",
  post_text_version = ")",
  overwrite = TRUE,
  display = FALSE
)
}
\arguments{
\item{file_name}{A character string, the file
name for the R script to update.}

\item{prior_text_package}{A character string, the
text preceding the package name.}

\item{prior_text_version}{A character string, the
text preceding the package's version number.}

\item{post_text_version}{A character string, the
text following the package's version number
(can just be the initial few characters).}

\item{overwrite}{Logical; if \code{TRUE} updates
the original R script by overwriting its content.}

\item{display}{Logical; if \code{TRUE} displays
the updated R script content to the console window.}
}
\value{
Either as a side effect overwrites the original
R script content or displays the updated content to the
console window.
}
\description{
Function to update the documentation in an R
script with a package's most up-to-date version
number.
}
