% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05_Statistics.R
\name{yeo_johnson}
\alias{yeo_johnson}
\title{\enumerate{
\item Yeo-Johnson Transform
}}
\usage{
yeo_johnson(y, lambda)
}
\arguments{
\item{y}{A numeric vector (values can be
positive, negative, or zero).}

\item{lambda}{A numeric value specifying
the transformation to apply.}
}
\value{
A numeric vector.
}
\description{
Transforms a numeric vector of values using
the Yeo-Johnson (2000) power transformation
family.
}
\examples{
# Example from page 958 of Yeo & Johnson (2000)
y <- c(
  6.1, -8.4, 1.0, 2.0, 0.7, 2.9, 3.5,
  5.1, 1.8, 3.6, 7.0, 3.0, 9.3, 7.5, -6.0
)
shapiro.test( y ) # Test of normality
y_transformed <- yeo_johnson(y, 1.305)
shapiro.test( y_transformed ) # Test of normality shows improvement

}
\references{
Yeo, I. K., & Johnson, R. A. (2000). A new family of power
transformations to improve normality or symmetry. Biometrika,
87 (4), 954-959. https://doi.org/10.1093/biomet/87.4.954
}
