% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05_Statistics.R
\name{yeo_johnson_transform}
\alias{yeo_johnson_transform}
\title{Determine Best Parameter for Yeo-Johnson Transform}
\usage{
yeo_johnson_transform(x, lower = -100, upper = 100)
}
\arguments{
\item{x}{A numeric vector (values can be
positive, negative, or zero).}

\item{lower}{The smallest value for the transformation
parameter to consider.}

\item{upper}{The highest value for the transformation
parameter to consider.}
}
\value{
A numeric vector, the transformed values of x.
}
\description{
Estimates and applies best-fitting Yeo-Johnson
transformation parameter via maximum likelihood
for a vector of numeric values.
}
\details{
The transformation parameter to use is estimated via
maximum likelihood using the \link{base:optimize} and
\link{stats:dnorm} functions.
}
\examples{
# Example from page 958 of Yeo & Johnson (2000)
x <- c(
  6.1, -8.4, 1.0, 2.0, 0.7, 2.9, 3.5,
  5.1, 1.8, 3.6, 7.0, 3.0, 9.3, 7.5, -6.0
)
shapiro.test( x ) # Test of normality
x_transformed <- yeo_johnson_transform(x)
# Extract results of maximum likelihood estimation
attributes(x_transformed)$mle_for_yeo_johnson
shapiro.test( x_transformed ) # Test of normality shows improvement

}
\references{
Yeo, I. K., & Johnson, R. A. (2000). A new family of power
transformations to improve normality or symmetry. Biometrika,
87 (4), 954-959. https://doi.org/10.1093/biomet/87.4.954
}
