% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arfpam-package.R
\docType{package}
\name{arfpam-package}
\alias{arfpam-package}
\title{Assorted R Functions for Processing, Analysis, and Modeling}
\description{
The \pkg{arfpam} package provides assorted
functions to help with a variety of common
tasks when processing, plotting, analyzing,
and modeling data.
}
