% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{copy_from_source}
\alias{copy_from_source}
\title{Copy Files From Source Folder}
\usage{
copy_from_source(
  path_to_source = "",
  renviron = "FOLDER_SOURCE",
  subfolder = "",
  new_folder = "Source"
)
}
\arguments{
\item{renviron}{An optional character
string, the environmental variable with
the path to the source folder.}

\item{subfolder}{An optional character
string, a subfolder to copy files from.
Partial matching of subfolder names is possible.}

\item{new_folder}{An optional character string,
the name of a subfolder to copy files to -
if it does not exist, the function will
create the subfolder.}

\item{path_to_folder}{A character string,
the absolute path to the source folder.}
}
\value{
As a side effect copies files and
folders.
}
\description{
Function to copy files from a subfolder
in a source folder to a new subfolder in
a user-defined source folder in the current
directory.
}
