% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{find_filename}
\alias{find_filename}
\title{...}
\usage{
find_filename(string, output = "Logical")
}
\arguments{
\item{string}{...}

\item{output}{...}
}
\value{
...
}
\description{
...
}
\examples{
# Forthcoming

}
