% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S04_Plotting.R
\name{hv_line}
\alias{hv_line}
\title{Draw Horizontal/Vertical Lines}
\usage{
hv_line(y = NULL, x = NULL, l = NULL, ...)
}
\arguments{
\item{y}{A vector with the y-axis positions
for horizontal lines.}

\item{x}{A vector with the x-axis positions
for vertical lines.}

\item{l}{...}

\item{...}{Additional arguments to be
passed to the \code{\link[graphics]{segments}}
function.}
}
\description{
...
}
\examples{
# Create a blank plot
blank_plot( xl, yl )

# Draw horizontal line
hv_line( y = .5, lty = 2 )
# Draw vertical line
hv_line( x = .5, lwd = 2 )

# Control width of horizontal line
hv_line( y = .25, l = c( .25, .75 ), col = 'blue' )
# Control height of vertical line
hv_line( x = .25, l = c( .25, .75 ), col = 'orange' )

}
