% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S02_Math.R
\encoding{UTF-8}
\name{logistic}
\alias{logistic}
\title{Logistic Function}
\usage{
logistic(x, na.rm = TRUE)
}
\arguments{
\item{x}{A set of values (\code{-Inf} ≤
x ≤ \code{Inf}).}

\item{na.rm}{Logical; if \code{TRUE}, removes \code{NA}
values first.}
}
\value{
Returns a set of values between 0 and 1.
}
\description{
This function applies the logistic function to
a set of values.
}
\examples{
round(logistic(c(0, -2.197225, 2.1972255, -Inf, Inf)), 6)
}
