% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{make_file_name}
\alias{make_file_name}
\title{Create Formatted File Name}
\usage{
make_file_name(
  description,
  extension,
  tag = NULL,
  number = NULL,
  file_date = NULL,
  additional = NULL,
  remove = FALSE
)
}
\arguments{
\item{description}{...}

\item{extension}{...}

\item{tag}{...}

\item{number}{...}

\item{file_date}{...}

\item{additional}{...}

\item{remove}{...}
}
\value{
...
}
\description{
...
}
\examples{
# Forthcoming

make_file_name( 'Example', 'RData' )
make_file_name( 'Example', 'pdf' )
make_file_name( 'Example', 'docx' )

make_file_name( 'Example', 'RData', tag = 'R', number = '02' )

}
