% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S04_Plotting.R
\name{plot_forest}
\alias{plot_forest}
\title{Create a Forest Plot}
\usage{
plot_forest(
  values,
  lower = NULL,
  upper = NULL,
  point_type = 19,
  point_color = "black",
  point_background = "white",
  text_size = 1,
  line_width = 2,
  line_color = "black",
  margin = c(3.5, 10, 2, 0.5),
  labels_y = NULL,
  labels_x = NULL,
  labels_position = -1,
  title_x = NULL,
  xlim = NULL,
  vert_grid = 0,
  vert_color = "black",
  vert_type = 1,
  horiz_grid = FALSE,
  horiz_color = "grey80",
  new = FALSE
)
}
\arguments{
\item{values}{Either a vector of estimates or
a data.frame/matrix with three columns
consisting of the estimates, the lower limits
the for error bars, and the upper limits for
the error bars).}

\item{lower}{A vector matching in length to
\code{values} with the lower limits for
the error bars.}

\item{upper}{A vector matching in length to
\code{values} with the upper limits for
the error bars.}

\item{point_type}{An integer vector specifying the
type of point to draw (see \link[graphics:points]{graphics::points}).}

\item{point_color}{A character vector specifying
the color(s) for the points.}

\item{point_background}{A character vector specifying
the color(s) for the point backgrounds.}

\item{text_size}{A numeric vector of three elements
specifying the size of the points to draw, the
size of the x and y-axis labels, and the size
of the x-axis title, respectively. If less than
three elements are given, elements are recycled
to produce the necessary length.}

\item{line_width}{A numeric vector of three elements,
the line widths for the error bars, the horizontal
grid lines, and the vertical grid lines, respectively.}

\item{line_color}{A character vector specifying
the color(s) for the error bars.}

\item{margin}{A numeric vector specifying the margin
sizes for the bottom, left, top, and right-hand sides
of the figure.}

\item{labels_y}{A character vector matching in length
to \code{values}, the labels for the estimates.}

\item{labels_x}{A numeric vector giving the x-axis
labels. If \code{NULL} the function attempts to
automatically create the labels.}

\item{labels_position}{A numeric vector of three
elements, the position at which to draw the
y-axis labels, the x-axis labels, and the x-axis
title, respectively.}

\item{title_x}{A character string, the x-axis title.}

\item{xlim}{A numeric vector with the lower and
upper limits for the x-axis plotting boundary.
If \code{NULL} the function computes it using
\link[base:range]{base::range}.}

\item{vert_grid}{A numeric vector giving the x-axis
positions for vertical grid lines. If \code{NULL}
no grid lines are drawn.}

\item{vert_color}{A character vector specifying
the color(s) for the vertical grid lines.}

\item{horiz_grid}{Logical; if \code{TRUE} horizontal
grid lines are added to separate the different
estimates visually.}

\item{horiz_color}{A character vector specifying
the color for the horizontal grid lines.}

\item{new}{Logical; if \code{TRUE} a new plotting window
is generated.}
}
\value{
A forest plot.
}
\description{
Generates a simple forest plot (defined here as
a plot of multiple estimates and their associated
error bars).
}
\examples{
# Example estimates and error bar limits
  means <- runif( 5, -.5, .5 )
  mat <- sapply( means, function(m) { rnorm( 100, m, 1 ) } )
  values <- sapply(
    1:ncol(mat), function(i) {
      x <- mat[, i]
      return( c( mean(x), mean(x) + sem(x)*qnorm( c(.025, .975) ) ) )
    }
  ) |> t()

# Example forest plot
plot_forest(
  values, labels_y = 'Column ' \%p\% 1:5,
  title_x = 'Estimated mean'
)

}
