% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05_Statistics.R
\name{standardize}
\alias{standardize}
\title{Standardize Columns in a Data Frame or Matrix}
\usage{
standardize(x, y = NULL, columns = NULL, raw = FALSE, scaling = FALSE)
}
\arguments{
\item{x}{A data frame or matrix of numeric values.}

\item{y}{A data frame or matrix of numeric values
(must have the same column names as \code{x}).}

\item{columns}{A character vector of column names,
the subset of columns in \code{x} to standardize.}

\item{raw}{Logical; if \code{TRUE}, uses the means and
standard deviations stored in each column's
attributes to restore scaled values to their
original raw values.}

\item{scaling}{Logical; if \code{TRUE} returns
the means and standard deviations for columns
stored in their attributes.}
}
\value{
A data frame or matrix with standardized
columns.
}
\description{
Function to standardize (mean-center and scale
by standard deviation resulting in a mean of 0
and standard deviation of 1) columns in a
matrix or data frame.
}
\examples{
# Create data frame
x_raw <- matrix( rnorm( 1000, 100, 15 ), 100, 10 )
colnames(x_raw) <- paste0( 'X', 1:10 )
x_raw <- data.frame( x_raw )

# Standardize columns
x <- standardize( x_raw )

# Create second data frame with same
# variables but new values
y <- matrix( rnorm( 50*10, 50, 15 ), 50, 10 )
colnames(y) <- paste0( 'X', 1:10 )
y <- data.frame( y )

# Scale columns of y based on means and
# standard deviations from x
y <- standardize( x, y )

# Undo scaling
all( round( x_raw, 4 ) == round( standardize( x, raw = TRUE ), 4 ) )

}
