% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05_Statistics.R
\name{stats_by_group}
\alias{stats_by_group}
\title{Compute Statistics by Group}
\usage{
stats_by_group(
  dtf,
  column,
  groupings = NULL,
  statistics = c("M", "SD"),
  method = "Student's T",
  categories = 1,
  width = 0.95,
  na.rm = TRUE
)
}
\arguments{
\item{dtf}{A data frame.}

\item{column}{A character string, the column in \code{dtf} to
compute statistics for.}

\item{groupings}{A character vector, the columns in \code{dtf}
to use as grouping factors (it is recommended that they all
be categorical variables).}

\item{statistics}{A character vector, the set of different
statistics to compute over groups.}

\item{method}{A character string, the type of method to use
when computing uncertainty intervals. Options include:
\code{"Student's T"} for means or
\code{"Beta-binomial"} for proportions.}

\item{categories}{An optional vector of elements to match over
when computing frequencies, proportions, or percentages.}

\item{width}{A numeric value between 0 and 1, the width for
uncertainty intervals.}

\item{na.rm}{A logical value; if \code{TRUE} removes
\code{NA} values.}
}
\value{
A data frame with separate rows for each combination
of grouping factors and separate columns for each statistic
to compute.
}
\description{
A function to compute assorted univariate statistics
for a specified variable in a data frame over desired
grouping factors.
}
\details{
Possible univariate statistics that can be computed:
\itemize{
\item \code{'N'} = Sample size;
\item \code{'M'} = Mean;
\item \code{'Md'} = Median;
\item \code{'SD'} = Standard deviation;
\item \code{'SE'} = Standard error of the mean;
\item \code{'C'} = Counts/frequencies;
\item \code{'Pr'} = Proportions;
\item \code{'Pe'} = Percentages;
\item \code{'Mn'} = Minimum;
\item \code{'Mx'} = Maximum;
\item \code{'Q1'} = First quartile;
\item \code{'Q3'} = Third quartile;
\item \code{'Q___'} = Quantile for specified percentage
(sub \code{'001'} to \code{'100'} for \code{'___'});
\item \code{'CNA'} Counts/frequencies for \code{NA} values;
\item \code{'PrNA'} Proportions for \code{NA} values;
\item \code{'PeNA'} Percentages for \code{NA} values.
}

Additionally, specifying \code{'UI'} in combination with the
argument \code{method} will compute the lower and upper limits
of a desired uncertainty interval. The width of the interval
can be controlled by the argument \code{width}.
}
\examples{
# Example data set
data("iris")

# Mean/SD for sepal length by species
iris |> stats_by_group( Sepal.Length, nq(Species) )

# Define categorical variable for long petals based on median split
iris$Long.Petal <- assign_by_interval(
  iris$Petal.Length, median(iris$Petal.Length), values = c( 'No', 'Yes' )
)
# Sample size, mean, and confidence intervals using Student's T
# distribution by species and whether petals are long
iris |> stats_by_group(
  Sepal.Length, nq( Species, Long.Petal ), nq( N, M, UI )
)

# Define categorical variable for long sepal based on median split
iris$Long.Sepal <- assign_by_interval(
  iris$Sepal.Length, median(iris$Sepal.Length), values = c( 'No', 'Yes' )
)
# Proportion and confidence intervals based on beta-binomial
# distribution for long sepal by long petal
iris |> stats_by_group(
  Long.Sepal, nq( Long.Petal ), nq( Pr, UI ),
  categories = 'Yes', method = 'Beta-binomial'
)

# Standard cut-offs for boxplots (min, Q1, median, Q3, max)
iris |> stats_by_group(
  Sepal.Length, nq( Species ), nq( Mn, Q1, Md, Q3, Mx )
)

# Custom quantiles
iris |> stats_by_group(
  Sepal.Length, nq( Species ), nq( Q005, Q025, Q050, Q075, Q095 )
)

# Missing values
iris$Sepal.Length[ c( 10, 30, 55:60, 120:123 ) ] <- NA
iris |> stats_by_group(
  Sepal.Length, nq( Species ), nq( CNA, PrNA, PeNA )
)

}
