% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05_Statistics.R
\name{summa}
\alias{summa}
\title{Flexible Formatted Summary Statistics}
\usage{
summa(
  x,
  syntax = "[[M]] ([[SD]])",
  categories = NULL,
  digits = NULL,
  na.rm = TRUE,
  pad = FALSE,
  f = NULL,
  ...
)
}
\arguments{
\item{x}{A vector of values.}

\item{syntax}{A character string with
identifiers in the form \code{[[.]]}
(where \code{.} can be a variety of
letter sets for different summary statistics) -
the function then substitutes the appropriate
computed value for the corresponding
identifier (see details for more information).}

\item{categories}{An optional vector of
elements to match in \code{x} when
computing frequencies, proportions,
or percentages.}

\item{digits}{Number of digits to round
summary statistics.}

\item{na.rm}{Logical; if \code{TRUE}
removes \code{NA} values from \code{x}.}

\item{pad}{Logical; if \code{TRUE} pads
values with 0 to all have a matching
number of decimal places.}

\item{f}{An optional user-defined function
that takes \code{x} as a first argument
and returns a vector of values.
The i-th outputted value will then
be substituted for the corresponding
identifier \code{[[i]]} (see examples).}

\item{...}{Additional arguments for the
user-defined function \code{f}.}
}
\value{
A character string.
}
\description{
A function that allows users to create
a nicely formatted character string
with summary statistics based on
user-supplied identifiers via a
simple, intuitive syntax.
}
\details{
This function provides some simple syntax
to allow users to write out a custom phrase for
reporting summary statistics.
The function then searches the input for
identifiers - once found, the function computes
the appropriate summary statistic and
substitutes the numeric result in place of
the given identifier.

For example, a user can provide the phrase:

\code{'Mean = [[M]]'},

and the function will then substitute the sample
mean of the vector \code{x} for the identifier
\code{[[M]]}.

Pre-defined identifiers are:
\itemize{
\item \code{[[N]]} = Sample size;
\item \code{[[M]]} = Mean;
\item \code{[[SD]]} = Standard deviation;
\item \code{[[SE]]} = Standard error of the mean;
\item \code{[[Mn]]} = Minimum;
\item \code{[[Q1]]} = 1st quartile;
\item \code{[[Md]]} = Median;
\item \code{[[Q3]]} = 2nd quartile;
\item \code{[[Mx]]} = Maximum;
\item \code{[[IQR]]} = Inter-quartile range;
\item \code{[[C]]} = Counts/frequencies;
\item \code{[[P]]} = Percent;
\item \code{[[Pr]]} = Proportion.
}

Users can also pass in a custom function \code{f}
that takes \code{x} as a first argument and
returns a vector of values. Then element \code{i}
from the outputted vector is substituted for
the identifier \code{[[i]]}.
}
\examples{
# Example using 'iris' data set
data("iris")
# Continuous variable - sepal length
x <- iris$Sepal.Length

# Mean and standard deviation
summa(x)
# Median and IQR
summa(x, "[[M]] ([[IQR]])")
# Pad to 2 decimal places
summa(x, "[[M]] ([[IQR]])", pad = TRUE)
# Mean (SD); N [min and max]
summa(x, "[[N]]; [[M]] ([[SD]]); " \%p\%
  "[[[Mn]], [[Q1]], [[Md]], [[Q3]], [[Mx]]]",
digits = 1
)

# Custom measures via user-defined function
# (e.g., bootstrapped confidence interval)
fnc <- function(x) {
  btstrp <- bootstrap(
    x,
    summary = function(y) quantile(y, c(.025, .975))
  )
  return(btstrp$summary)
}
summa(x, "[[M]] ([[SE]]) [[[1]] to [[2]]]",
  f = fnc
)

# Example using 'mtcars' data set
# Categorical variable - # of forward gears
data("mtcars")
x <- mtcars$gear

# Percent and counts for 3 forward gears
summa(x == 3, "[[P]]\% ([[C]] out of [[N]])")
# Percent and counts for 4 or 5 forward gears
summa(x, "[[P]]\% ([[C]] out of [[N]])",
  categories = c(4, 5)
)
}
