% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{templates}
\alias{templates}
\title{Create Templates for Annotations and Code}
\usage{
templates(type)
}
\arguments{
\item{type}{The type of template to return,
with options for...
\itemize{
\item 'Function (function documentation);
\item 'Header' (header for a R script);
\item 'Progress' (progress bar for a loop);
\item 'Segment' (code segment to conditionally run);
\item 'Loop' (a \code{for} loop statement);
\item 'Conditional' (a \code{if} statement).
}}
}
\description{
Outputs a template of common annotations
and code segments to the console for
easy copying and pasting.
}
\examples{
# List of possible inputs to argument
# 'type' for each template
templates()

# Function documentation
templates("Function")

# Header for R script
templates("Header")

# Progress bar for loop
templates("Progress")

# Code segment in a script
templates("Segment")
}
