% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03_Utilities.R
\name{templates}
\alias{templates}
\title{Create Templates for Annotations and Code}
\usage{
templates(type = NULL, val = NULL)
}
\arguments{
\item{type}{The type of template to return,
with options for...
\itemize{
\item 'Function (function documentation);
\item 'Header' (header for a R script);
\item 'Progress' (progress bar for a loop);
\item 'Loop' (a \code{for} loop statement);
\item 'Conditional' (a \code{if} statement);
\item 'recode' (values for the 'dplyr'
function \code{\link[dplyr]{recode}}).
}}

\item{val}{An optional character vector to be used
with \code{type = 'recode'}.}
}
\description{
Outputs a template of common annotations
and code segments to the console for
easy copying and pasting.
}
\examples{
# List of possible inputs to argument
# 'type' for each template
templates()

# Function documentation
templates("Function")

# Header for R script
templates("Header")

# Progress bar for loop
templates("Progress")

# Loop
templates("Loop")

# If ... else statement
templates("Conditional")

# Roxygen2 function documentation
templates("roxygen")

# HTML internal links
templates("html_links")

# Example table for specifying
# nomenclature and a glossary
# in a .Rmd file
templates("rmd_glossary")

}
